gastos_bot/requirements.txt
python-telegram-bot>=20.0
gspread>=6.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
faster-whisper>=1.0.0
google-generativeai>=0.8.0
